reorganizes tests in commons, export only shared stuff in test jar
