+++ apis/sgv2-quarkus-common/pom.xml
@@ -166,6 +166,11 @@
             <goals>
               <goal>test-jar</goal>
             </goals>
+            <configuration>
+              <excludes>
+                <exclude>**/io/stargate/sgv2/api/common/*</exclude>
+              </excludes>
+            </configuration>
           </execution>
         </executions>
       </plugin>
