+++ apis/sgv2-quarkus-common/src/test/java/io/stargate/sgv2/api/common/properties/datastore/configuration/BridgeDataStorePropertiesRetryTest.java
@@ -14,9 +14,9 @@ import io.quarkus.grpc.GrpcClient;
 import io.quarkus.test.junit.QuarkusTest;
 import io.stargate.bridge.proto.Schema;
 import io.stargate.bridge.proto.StargateBridgeGrpc;
-import io.stargate.sgv2.api.common.BridgeTest;
 import io.stargate.sgv2.api.common.config.DataStoreConfig;
 import io.stargate.sgv2.api.common.properties.datastore.DataStoreProperties;
+import io.stargate.sgv2.common.BridgeTest;
 import java.util.concurrent.atomic.AtomicInteger;
 import javax.inject.Inject;
 import org.junit.jupiter.api.Test;
