+++ apis/sgv2-quarkus-common/src/test/java/io/stargate/sgv2/api/common/schema/SchemaManagerTest.java
@@ -43,10 +43,10 @@ import io.stargate.bridge.grpc.Values;
 import io.stargate.bridge.proto.QueryOuterClass;
 import io.stargate.bridge.proto.Schema;
 import io.stargate.bridge.proto.StargateBridge;
-import io.stargate.sgv2.api.common.BridgeTest;
 import io.stargate.sgv2.api.common.StargateRequestInfo;
 import io.stargate.sgv2.api.common.grpc.UnauthorizedKeyspaceException;
 import io.stargate.sgv2.api.common.grpc.UnauthorizedTableException;
+import io.stargate.sgv2.common.BridgeTest;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
